[Refactor] 싱글턴 변수명 변경, 코딩 컨벤션 적용
싱글턴 변수명: shared
상속되지 않는 클래스에 final 키워드 추가
