+++ FlyingPopcornApp/FlyingPopcornApp/App/SceneDelegate.swift
@@ -24,12 +24,12 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 //        window?.rootViewController = tabBarViewController
         
         /// 저장된 유저 체크
-        if let userEmail = UserDefaultsHelper.userDefaultsHelper.getCurrentUser() {
+        if let userEmail = UserDefaultsHelper.shared.getCurrentUser() {
             // 저장된 사용자 정보로 UserData 업데이트
-            UserDefaultsHelper.userDefaultsHelper.loadUserData(email: userEmail)
+            UserDefaultsHelper.shared.loadUserData(email: userEmail)
             
             // 데이터 체크
-            print("family-name: \(String(describing: UserData.loginedUser.familyName)), name: \(String(describing: UserData.loginedUser.name)), email: \(String(describing: UserData.loginedUser.email)), tickets: \(UserData.loginedUser.tickets)")
+            print("family-name: \(String(describing: UserData.shared.familyName)), name: \(String(describing: UserData.shared.name)), email: \(String(describing: UserData.shared.email)), tickets: \(UserData.shared.tickets)")
             
             // 자동 로그인 -> 메인 화면으로 이동
             let tabBarViewController = BottomTabbarViewController(movieNetwork: movieNetwork)
