--- FlyingPopcornApp/FlyingPopcornApp/Sources/Claire/Delegate/SwitchViewDelegate.swift
@@ -1,15 +1,0 @@
-//
-//  SwitchViewDelegate.swift
-//  FlyingPopcornApp
-//
-//  Created by Claire on 12/19/24.
-//
-
-protocol SigninViewDelegate: AnyObject {
-    func didTapSignupButton()
-    func moveToMain()
-}
-
-protocol SignupViewDelegate: AnyObject {
-    func didTapSigninButton()
-}
