+++ FlyingPopcornApp/FlyingPopcornApp/Sources/Claire/View/MainSolidButton.swift
@@ -4,7 +4,9 @@
 //
 //  Created by CHYUN on 12/16/24.
 //
+
 import UIKit
+
 import SnapKit
 
 final class MainSolidButton: UIButton {
