+++ FlyingPopcornApp/FlyingPopcornApp/Sources/Claire/View/SignSwitchLabel.swift
@@ -4,6 +4,7 @@
 //
 //  Created by CHYUN on 12/17/24.
 //
+
 import UIKit
 
 final class SignSwitchLabel: UIView {
