+++ FlyingPopcornApp/FlyingPopcornApp/Sources/Claire/View/StrokeTextField.swift
@@ -4,6 +4,7 @@
 //
 //  Created by CHYUN on 12/16/24.
 //
+
 import UIKit
 
 class StrokeTextField: UITextField {
