--- FlyingPopcornApp/FlyingPopcornApp/Sources/Hamin/Controllers/MainViewController.swift
@@ -1,19 +1,0 @@
-//
-//  MainViewController.swift
-//  FlyingPopcornApp
-//
-//  Created by t0000-m0112 on 2024-12-13.
-//
-
-import UIKit
-
-class MainViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
