+++ FlyingPopcornApp/FlyingPopcornApp/Sources/Sia/DetailView/MovieDetailView.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+
 import SnapKit
 import Then
