+++ FlyingPopcornApp/FlyingPopcornApp/Sources/Sia/DetailView/RateView.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+
 import SnapKit
 import Then
