+++ FlyingPopcornApp/FlyingPopcornApp/Utilities/Extensions/UIImage+Extension.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+
 import Kingfisher
 
 extension UIImageView {
